+++ demos/chinook/src/main/java/is/codion/demos/chinook/ui/InvoiceTablePanel.java
@@ -40,7 +40,7 @@ public final class InvoiceTablePanel extends EntityTablePanel {
 						// see InvoiceLineEditModel, so we don't want it to be editable via the popup menu.
 						.editable(attributes -> attributes.remove(Invoice.TOTAL))
 						// The factory providing our custom condition panel.
-						.conditionPanelFactory(new InvoiceConditionPanelFactory(tableModel))
+						.conditionPanel(new InvoiceConditionPanelFactory(tableModel))
 						// Start with the SIMPLE condition panel view.
 						.conditionView(SIMPLE));
 	}